Add abstract_data option
