+++ experimental/tools/unit-generators/smv/generators/arith/addi.py
@@ -1,13 +1,16 @@
-from generators.support.binary_op_handshake_manager import generate_binary_op_handshake_manager
-from generators.support.arith_op_headers import generate_binary_op_header
+from generators.support.arith_utils import *
 from generators.support.utils import *
 
 
 def generate_addi(name, params):
   latency = params[ATTR_LATENCY]
   data_type = SmvScalarType(params[ATTR_DATA_TYPE])
+  abstract_data = params[ATTR_ABSTRACT_DATA]
 
-  return _generate_addi(name, latency, data_type)
+  if abstract_data:
+    return generate_abstract_binary_op(name, latency, data_type)
+  else:
+    return _generate_addi(name, latency, data_type)
 
 
 def _generate_addi(name, latency, data_type):
