+++ experimental/tools/unit-generators/smv/generators/arith/negf.py
@@ -1,19 +1,13 @@
-from generators.support.delay_buffer import generate_delay_buffer
-from generators.support.arith_op_headers import generate_unanary_op_header
+from generators.support.arith_utils import generate_abstract_unary_op
 from generators.support.utils import *
 
 
 def generate_negf(name, params):
   latency = params[ATTR_LATENCY]
   data_type = SmvScalarType(params[ATTR_DATA_TYPE])
+  abstract_data = params[ATTR_ABSTRACT_DATA]
 
-  return _generate_negf(name, latency, data_type)
-
-
-def _generate_negf(name, latency, data_type):
-  return f"""
-{generate_unanary_op_header(name)}
-  DEFINE outs := ins;
-  
-  {generate_delay_buffer(f"{name}__delay_buffer", {"latency": latency})}
-"""
+  if abstract_data:
+    return generate_abstract_unary_op(name, latency, data_type)
+  else:
+    raise ValueError("Floating point operations support abstract data only")
