+++ experimental/tools/unit-generators/smv/generators/arith/truncf.py
@@ -1,5 +1,4 @@
-from generators.support.delay_buffer import generate_delay_buffer
-from generators.support.arith_op_headers import generate_unanary_op_header
+from generators.support.arith_utils import generate_abstract_unary_op
 from generators.support.utils import *
 
 
@@ -7,14 +6,9 @@ def generate_truncf(name, params):
   latency = params[ATTR_LATENCY]
   input_type = SmvScalarType(params[ATTR_INPUT_TYPE])
   output_type = SmvScalarType(params[ATTR_OUTPUT_TYPE])
+  abstract_data = params[ATTR_ABSTRACT_DATA]
 
-  return _generate_truncf(name, latency, input_type, output_type)
-
-
-def _generate_truncf(name, latency, input_type, output_type):
-  return f"""
-{generate_unanary_op_header(name)}
-  DEFINE outs := ins;
-  
-  {generate_delay_buffer(f"{name}__delay_buffer", {"latency": latency})}
-"""
+  if abstract_data:
+    return generate_abstract_unary_op(name, latency, output_type)
+  else:
+    raise ValueError("Floating point operations support abstract data only")
