--- experimental/tools/unit-generators/smv/generators/support/arith_op_headers.py
@@ -1,21 +1,0 @@
-def generate_binary_op_header(name):
-  return f"""
-MODULE {name}(lhs, lhs_valid, rhs, rhs_valid, outs_ready)
-  VAR inner_handshake_manager : {name}__handshake_manager(lhs_valid, rhs_valid, outs_ready);
-
-  // output
-  DEFINE lhs_ready := inner_handshake_manager.lhs_ready;
-  DEFINE rhs_ready := inner_handshake_manager.rhs_ready;
-  DEFINE outs_valid := inner_handshake_manager.outs_valid;
-"""
-
-
-def generate_unanary_op_header(name):
-  return f"""
-MODULE {name}(ins, ins_valid, outs_ready)
-  VAR inner_delay_buffer : {name}__delay_buffer(ins_valid, outs_ready);
-
-  // output
-  DEFINE ins_ready := inner_delay_buffer.ins_ready;
-  DEFINE outs_valid := inner_delay_buffer.outs_valid;
-"""
