+++ experimental/tools/unit-generators/smv/generators/support/utils.py
@@ -8,6 +8,7 @@ ATTR_VALUE = "value"
 ATTR_PORT_TYPES = "port_types"
 ATTR_LATENCY = "latency"
 ATTR_PREDICATE = "predicate"
+ATTR_ABSTRACT_DATA = "abstract_data"
 
 
 class SmvScalarType:
@@ -65,6 +66,7 @@ class SmvScalarType:
 
 
 HANDSHAKE_CONTROL_TYPE = SmvScalarType("!handshake.control<>")
+
 
 def TEHB_BUFFER_PARAMS(data_type):
   return {ATTR_SLOTS: 1, ATTR_TIMING: "R: 1", ATTR_PORT_TYPES: {"outs": data_type.mlir_type}}
